Mejoras UI, formularios, CLP y sin decimales
- Dashboard: estilos en app.css, datos dinámicos (semanas totales)
- Formularios: estilos centralizados, sin inline styles
- Ranking: filter-form con estilos, inputs con appearance:none
- Configuración: textarea para JSON complejos, filtro setting_is_json
- Sin decimales: IntegerField valor hora, montos enteros, display %.0f
- Peso chileno (CLP): filtro clp con formato $1.000 CLP
- Alertas: estilos movidos a app.css

Made-with: Cursor

+++ alfajor/blueprints/employees/forms.py
@@ -1,7 +1,7 @@
 """Formularios empleados."""
 
 from flask_wtf import FlaskForm
-from wtforms import StringField, FloatField, SelectField, BooleanField, TextAreaField
+from wtforms import StringField, IntegerField, SelectField, BooleanField, TextAreaField
 from wtforms.validators import DataRequired, Optional, NumberRange
 
 
@@ -17,7 +17,7 @@ class EmployeeForm(FlaskForm):
         ("cocina", "Cocina"),
         ("runner", "Runner"),
     ], validators=[Optional()])
-    hourly_rate = FloatField("Valor hora", validators=[DataRequired(), NumberRange(min=0)])
+    hourly_rate = IntegerField("Valor hora", validators=[DataRequired(), NumberRange(min=0)])
     status = SelectField("Estado", choices=[
         ("ACTIVO", "Activo"),
         ("INACTIVO", "Inactivo"),
